Re-organize Static3 Searcher, Generator and Util
