+++ rng_tools/src/generators/gen3/mod.rs
@@ -3,6 +3,7 @@ mod feebas_sid;
 mod seed;
 mod stationary;
 mod tidsid;
+mod util;
 
 pub use egg::*;
 pub use feebas_sid::*;
